Update the chat app integration and functions
