+++ client/src/pages/api-routes/APIRoutes.js
@@ -2,13 +2,13 @@ export const host = "http://localhost:4001";
 export const registerRoute = `${host}/api/auth/register`;
 export const loginRoute = `${host}/api/auth/login`;
 export const logoutRoute = `${host}/api/auth/logout`;
-export const allUsersRoute = `${host}/api/auth/allusers`;
+export const getAllUsersRoute = `${host}/api/auth/allusers`;
 export const accountSettingsRoute = `${host}/api/auth/accountsettings/`;
 export const accountDeleteRoute = `${host}/api/auth/accountdelete/`;
 export const sendMessageRoute = `${host}/api/messages/addmsg`;
 export const recieveMessageRoute = `${host}/api/messages/getmsg`;
 export const setAvatarRoute = `${host}/api/auth/setavatar`;
-export const UserRoute = `${host}/api/auth/user`;
+export const getUserRoute = `${host}/api/auth/user`;
 export const createChatRoute = `${host}/chat`;
 export const userChatRoute = `${host}/chat/userchat`;
 export const findChathatRoute = `${host}/chat/find/:firstId/:secondId`;
