+++ client/src/pages/chat/SidebarChat.js
@@ -2,8 +2,14 @@ import React, { useState, useEffect } from "react";
 import "./SidebarChat.css";
 import { Avatar } from "@mui/material";
 
-function SidebarChat({ addNewChat, username }) {
+function SidebarChat({ addNewChat, contacts, changeChat }) {
   const [currentSelected, setCurrentSelected] = useState(undefined);
+
+  const changeCurrentChat = (index, contact) => {
+    setCurrentSelected(index);
+    changeChat(contact);
+  };
+
   const createChat = () => {
     const userName = prompt("Please enter a user to chat with");
 
@@ -13,26 +19,26 @@ function SidebarChat({ addNewChat, username }) {
   };
 
   return !addNewChat ? (
-    <div className="sidebarChat">
-      <div className="sidebarChat_info">
-        {/*contacts.map((contact, index) => {
-          return (
-            <div
-              key={index}
-              className={`contact ${
-                index === currentSelected ? "selected" : ""
-              }`} /*
+    <div className="contacts">
+      {contacts.map((contact, index) => {
+        return (
+          <div
+            key={index}
+            className={`contact ${index === currentSelected ? "selected" : ""}`}
             onClick={() => changeCurrentChat(index, contact)}
-            >
-              <div className="username">
-                <h2>{contact.username}</h2>
-              </div>
+          >
+            {contact.avatarImage ? (
+              <img src={contact.avatarImage} alt="" />
+            ) : (
+              <Avatar />
+            )}
+            <div className="username">
+              <h2>{contact.username}</h2>
+              <p>This is the last message</p>
             </div>
-          );
-        })*/}
-        <h2>{username}</h2>
-        <p>This is the last message</p>
-      </div>
+          </div>
+        );
+      })}
     </div>
   ) : (
     <div onClick={createChat} className="sidebarChat">
