+++ client/src/pages/chat/Welcome.css
@@ -1,0 +1,20 @@
+.welcome {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  background-color: white;
+  flex: 0.65;
+}
+
+.welcome > img {
+  height: 20rem;
+}
+
+.welcome > h1 > span {
+  color: #235542;
+}
+
+.welcome > h2 {
+  font-size: 1.5rem;
+}
