+++ server/controllers/users.controller.js
@@ -45,6 +45,8 @@ module.exports.getUser = async (req, res, next) => {
     const users = await User.find({ _id: { $in: req.params.id } }).select([
       "email",
       "username",
+      "fullname",
+      "avatarImage",
       "_id",
     ]);
     return res.json(users);
@@ -59,6 +61,8 @@ module.exports.getAllUsers = async (req, res, next) => {
     const users = await User.find({ _id: { $ne: userId } }).select([
       "email",
       "username",
+      "fullname",
+      "avatarImage",
       "_id",
     ]);
     return res.json(users);
